Conflicts:
	common/src/addrtranslate-sysv.h
	dyninstAPI/src/BPatch.C
	dyninstAPI/src/aix.C
	dyninstAPI/src/emit-x86.C
	dyninstAPI/src/instPoint.C
	dyninstAPI/src/linux-x86.C
	dyninstAPI/src/linux.C
	dyninstAPI/src/process.C
	dyninstAPI/src/signalhandler.C
	dyninstAPI/src/solaris.C
	proccontrol/src/process.C
	stackwalk/h/procstate.h
	stackwalk/src/catamount-swk.C
	stackwalk/src/linux-swk.h
	stackwalk/src/procstate.C

+++ dyninstAPI/src/function.C
@@ -1987,7 +1987,8 @@ bblInstance * bblInstance::getTargetBBL() {
                 }
             }
             if (hlTarget == NULL) {
-                fprintf(stderr, "targets:%d out_edges:%d src:0x%lx->0x%lx trg:0x%lx->0x%lx\n", targets.size(), out_edges.size(), (*eit)->src()->start(), (*eit)->src()->end(), (*eit)->trg()->start(), (*eit)->trg()->end());
+                fprintf(stderr, "targets:%d out_edges:%d src:0x%lx->0x%lx trg:0x%lx->0x%lx\n", 
+                        (int)targets.size(), (int)out_edges.size(), (*eit)->src()->start(), (*eit)->src()->end(), (*eit)->trg()->start(), (*eit)->trg()->end());
             }
 
             assert(hlTarget != NULL);
